Move json_rpc_service::Background to its own module
