+++ light-base/src/json_rpc_service/background/getters.rs
